compiler warnings, small fixes

+++ simulation/simulation/ResultStruct.h
@@ -4,12 +4,12 @@
 struct SimulationResult
 {
 	SimulationResult() : 
-		reliability(1.0), meanAvailability(1.0), mttf(-1.0), nRounds(0), nFailures(0), duration(0) {};
+		reliability(1.0), meanAvailability(1.0), mttf(-1.0), nRounds(0), nFailures(0), duration(0.0) {};
 	
 	long double reliability;
 	long double meanAvailability;
 	long double mttf;
 	unsigned int nRounds;
 	unsigned int nFailures;
-	unsigned int duration;
+	double duration;
 };
